+++ GPTifier/include/datadir.hpp
