+++ GPTifier/src/configs.cpp
@@ -1,5 +1,5 @@
 #include "configs.hpp"
-#include "data_dir.hpp"
+#include "datadir.hpp"
 
 #include <filesystem>
 #include <stdexcept>
